requirements.txt
asgiref==3.6.0
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
dj-database-url==1.3.0
Django==4.2
django-admin-honeypot @ git+https://github.com/dmpayton/django-admin-honeypot.git@a840496d183df89965eeb33c9a5dd9ea3919dfff
django-cors-headers==3.14.0
django-filter==23.1
django-heroku==0.3.1
django-ipware==5.0.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
drf-yasg==1.21.5
Faker==18.4.0
gunicorn==20.0.4
i18n==0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
py==1.11.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
validate-docbr==1.10.0
whitenoise==6.4.0
